Move LoadEstimator code to folder
