+++ inc/Loading/ConstLoadEstimator.h
@@ -24,8 +24,8 @@
 #ifndef PCOE_CONST_LOAD_EST_H
 #define PCOE_CONST_LOAD_EST_H
 
-#include "LoadEstimator.h"
 #include "ConfigMap.h"
+#include "Loading/LoadEstimator.h"
 
 namespace PCOE {
     class ConstLoadEstimator : public LoadEstimator {
