+++ inc/Loading/LoadEstimatorFactory.h
@@ -4,10 +4,10 @@
 #ifndef PCOE_LOADESTIMATORFACTORY_H
 #define PCOE_LOADESTIMATORFACTORY_H
 
-#include "ConstLoadEstimator.h"
 #include "Factory.h"
-#include "LoadEstimator.h"
-#include "MovingAverageLoadEstimator.h"
+#include "Loading/ConstLoadEstimator.h"
+#include "Loading/LoadEstimator.h"
+#include "Loading/MovingAverageLoadEstimator.h"
 #include "Singleton.h"
 
 namespace PCOE {
