+++ src/Loading/ConstLoadEstimator.cpp
@@ -23,7 +23,7 @@
 
 #include <random>
 
-#include "ConstLoadEstimator.h"
+#include "Loading/ConstLoadEstimator.h"
 #include "ThreadSafeLog.h"
 
 namespace PCOE {
