+++ src/Loading/LoadEstimator.cpp
@@ -1,4 +1,4 @@
-#include "LoadEstimator.h"
+#include "Loading/LoadEstimator.h"
 
 namespace PCOE {
     void LoadEstimator::setNSamples(const unsigned int) {
