+++ src/ModelBasedPrognoser.cpp
@@ -27,7 +27,7 @@
 #include <iostream>
 
 #include "ConfigMap.h"
-#include "LoadEstimatorFactory.h"
+#include "Loading/LoadEstimatorFactory.h"
 #include "ModelBasedPrognoser.h"
 #include "Observers/ObserverFactory.h"
 #include "Predictors/PredictorFactory.h"
@@ -51,7 +51,7 @@ namespace PCOE {
         : Prognoser(configMap), initialized(false) {
         // Check for required config parameters
         requireKeys(configMap,
-            {MODEL_KEY, OBSERVER_KEY, PREDICTOR_KEY, NUMSAMPLES_KEY, HORIZON_KEY});
+                    {MODEL_KEY, OBSERVER_KEY, PREDICTOR_KEY, NUMSAMPLES_KEY, HORIZON_KEY});
         /// TODO(CT): Move Model, Predictor subkeys into Model/Predictor constructor
 
         // Create Model
@@ -62,7 +62,8 @@ namespace PCOE {
         // Create Observer
         log.WriteLine(LOG_DEBUG, moduleName, "Creating Observer");
         ObserverFactory& pObserverFactory = ObserverFactory::instance();
-        observer = pObserverFactory.Create(configMap.getVector(OBSERVER_KEY)[0], model.get(), configMap);
+        observer =
+            pObserverFactory.Create(configMap.getVector(OBSERVER_KEY)[0], model.get(), configMap);
 
         // Create Load Estimator
         log.WriteLine(LOG_DEBUG, moduleName, "Creating Load Estimator");
@@ -95,8 +96,7 @@ namespace PCOE {
         loadEstimator->setModel(model.get());
 
         // Set configuration parameters
-        unsigned int numSamples =
-            static_cast<unsigned int>(configMap.getUInt64(NUMSAMPLES_KEY));
+        unsigned int numSamples = static_cast<unsigned int>(configMap.getUInt64(NUMSAMPLES_KEY));
         unsigned int horizon = static_cast<unsigned int>(configMap.getUInt64(HORIZON_KEY));
 
         // @todo: Refactor for event-driven architecture
